Tests for NotionDatabaseManager
