--- test/test/fixtures/NotionConverter/outputs/NotionConverter/output.html
@@ -1,1 +1,0 @@
-<p class="">some text</p><details><summary class="">What is love?</summary><div><p class="">Baby don’t hurt me…</p></div></details><ul><li class="">Oranges</li></ul><p class="">some text</p><ol><li class="">Lakers</li></ol><p class="">some text</p><p class=""></p><ul><li class="">Apples</li></ul><ol><li class="">Warriors</li></ol><ol><li class="">Mavericks</li></ol><p class=""></p><p class=""><strong>Bold </strong><em><strong>Italic </strong></em><u><em><strong>Underline </strong></em></u><u><del><em><strong>StrikeThrough</strong></em></del></u><u><em><strong> </strong></em></u><code><u><em><strong>Code </strong></em></u></code><span class="notion-yellow-background"><code><u><em><strong>Color</strong></em></u></code></span></p><p class=""></p><h1>Heading 1</h1><figure><video src="https://www.youtube.com/watch?v=dQw4w9WgXcQ" controls></video><figcaption></figcaption></figure><p class=""></p><p class=""></p><blockquote class=""><em><strong>“One must have chaos in one's self to give birth to a dancing star</strong></em><strong>.”</strong> -Nietzsche</blockquote><p class=""></p><p class=""></p><figure><img src="https://upload.wikimedia.org/wikipedia/commons/thumb/2/20/YouTube_2024.svg/500px-YouTube_2024.svg.png"><figcaption>Youtube Logo</figcaption></figure><figure><img src="https://prod-files-secure.s3.us-west-2.amazonaws.com/4c6329c4-ddf4-4a16-8065-9bb204231c5d/32632a86-ac65-4998-a785-446e7160c31b/logo.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&#x26;X-Amz-Content-Sha256=UNSIGNED-PAYLOAD&#x26;X-Amz-Credential=ASIAZI2LB4664GLICZFU%2F20250518%2Fus-west-2%2Fs3%2Faws4_request&#x26;X-Amz-Date=20250518T052450Z&#x26;X-Amz-Expires=3600&#x26;X-Amz-Security-Token=IQoJb3JpZ2luX2VjELH%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEaCXVzLXdlc3QtMiJHMEUCIE8E4JnC5eYhdlvvlOj0WHJfxB9VCwZlZb5q5hH%2BdQcBAiEApcm6Xatg%2BkJrGd5CPHjlVIQcwMArca2ZFsAXorNvctsq%2FwMIahAAGgw2Mzc0MjMxODM4MDUiDHSkNJgca%2BHUN0%2FSYircAyR7XymXwImZJLzSxFaJ9iRAk2fx8sjbPrlSy7OAHkj%2Fw8emablnhwiPE7N7vKJvKJvDFY%2FylE64jQw51Ijj8rX0vYVp2txGhDsmGTKS%2FIahHA1m%2F7mTF5xQ1i5Dc8zqFzYj37Gbs91w6%2FliidHyFSILly%2FT%2FEjUdrLak4DoomH2cDTCytPksgQKI8LOk00BaexEbfc7k5J2AqHTRPH9iRHuVJq3oPQqxIGypGTHgEa%2FhuzMCOrbWkjw4FNvwh0V3koSgNHkw2YbQPYvAvAdHSSOqhmvrZ5c6QNVqIkA8R32p2Wk6R44RozXzCQGUCMUaDZxP2XPCRxqx4tkxY%2F2W4ozvpJAtHnDgnldKad9GsxNx4LJrNfTQizelENQGBVBU32S%2F3oZ4hNx%2FfO%2B%2F8YP6kV0Ai6DlRx6cJhSsEAJpVrIdiDW7051WDp%2F%2BZuF4Sc5a4LXwNa8l8QFW7Hr91TkEnLrXK2LpziBIj6%2Fo5l38vtdt5dt6aaY64fgRsNHOHjQMzhcrKhWkSM2eYUpP1ghbSii%2FtzarIzLOygKntlVJJNS20xaJ4zmcoBKetekGgi8NPXWAIfW11PkaOmtti57TS9SRvBbD9kRrIjayAZz0QSqBvFHe6TILITTwdOoMOvVpMEGOqUBZ8hq%2FR0tD6mTbnHq7ADv1WaZjNmSPlSXT6H0cjufmhKh3RpxmwlTHy21ZhI2ZiLYs48dTLQLRMxIlwhUNc3riIDu23xYP1lYYrmMa8ODMwOH2GKXoiRSs0ebcE6QQGgYtrnHrOhAucygMVflrplW7dxd6DXN95V37%2F25rTBFtxcw51lk%2Bj1bkVtRAt2lq9yLkXQ%2FZbQ7SSEvznrsDXVmXEzRuGkP&#x26;X-Amz-Signature=dac2625c1fd3e142149c1430793412e6f3ba7cda4baba22bd802b0262779a83b&#x26;X-Amz-SignedHeaders=host&#x26;x-id=GetObject"><figcaption>Rosette logo</figcaption></figure><table><tbody><tr><td>Sr. No</td><td>Name</td><td>Apples</td></tr><tr><td>2</td><td>John</td><td><em><strong>10</strong></em></td></tr><tr><td>1</td><td>Alice</td><td><em><strong>6</strong></em></td></tr></tbody></table><p class=""><a data-notion-page-mention="126744b6-a7b9-433a-bf8a-78db40eb004e">Untitled</a> </p><h3>Heading 3</h3><code class="language-math">x^n + y^n = z^n</code><ul><li class="">Mangoes</li></ul><p class=""></p><h2>Heading 2</h2><pre><code class="language-javascript">console.log("Hello World!");</code></pre>
