--- test/test/outputs/full_page_with_all_elements.html
@@ -1,1 +1,0 @@
-<h1>Heading 1</h1><p class="">some text</p><h2>Heading 2</h2><p class="">some text</p><h3>Heading 3</h3><p class="">some text</p><figure><img src="https://prod-files-secure.s3.us-west-2.amazonaws.com/4c6329c4-ddf4-4a16-8065-9bb204231c5d/32632a86-ac65-4998-a785-446e7160c31b/logo.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&#x26;X-Amz-Content-Sha256=UNSIGNED-PAYLOAD&#x26;X-Amz-Credential=ASIAZI2LB4666SHTHLUW%2F20250517%2Fus-west-2%2Fs3%2Faws4_request&#x26;X-Amz-Date=20250517T092951Z&#x26;X-Amz-Expires=3600&#x26;X-Amz-Security-Token=IQoJb3JpZ2luX2VjEJ%2F%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEaCXVzLXdlc3QtMiJHMEUCIE2YynnXZt4CO7WNGSoiKh1FhM1IX0Oe%2FCtZqsTOQZV2AiEA%2FoiTobzoFNaIgJ0raaU%2BlTbWAvXgaNgrGU5eUT8WAfAq%2FwMIWBAAGgw2Mzc0MjMxODM4MDUiDOL6R8FsieKZgT5wBCrcA%2FvkzmM70%2BNq71VSKtm2A2npspPdbN2dleLWG3y3qPsX1ytc3igMS%2FAlkJ5k8K8fJORe0pem2EBxVER%2FVs9FP%2FE5abP7VpDvLGD%2F2E7PdmdAR1kj%2BXoLuWiAJqbpzFJI1IEvOp7fG6rEndW2GH2WNCL5eDtTRqSnVvar7iSKjw79ItG7GRgNL07MNsFRq8ogc2jCV6rErWZXtnKMcMQx9pCSq5SYtDaQiM%2FQM0OaSwh6dq4P26U2cBDSvZDrIrY5e%2BdeUvrCp%2F0dWHHSjG5%2BVJlE3Kt8qz4rWuqoEIprRrRVC8gXdKSIM6rb3w7b1RJGwwapbWR8Ajp%2Fe%2F4W9K4ebSsNZyR7rgBCPeYE8p0%2FW4FeV5vPujAU5YkW4Cs2%2Bh9PI8lADQkMxl%2Btk285qmNonofI3Q0V2ojvl8htIzEOnndrnAm3x3OvpXvWJ7F97XF%2FeDy50QJKLTS44kJc1rat6zuOzj8KlATf%2Fb3ovmS7yrNjvuYz62gl2HEi00KMrJVe7M5tAdLB%2BJ2PmYGUS%2BcYAhygRwwrUKWteur6jK306NNClKBt%2BfrYfG6JI6L92HaPc0bTdI3qfRmnoHO3bKI9kPKOBWOMuBIl6UMxbX9QT%2FxN51ys6PoSLb4wISimMI7goMEGOqUB8PqH5mJqXx44nRuWq93MhdV0iCDd5PgHr0xPgW3JChrPhtcJ2xbwqH8tHkPusDBUdt8MS%2BOJkcTLCED9Sf6ihHjtlwrVSqApGdm62bqLYoJPWo%2FnIfgiJy%2FASuPcxq3QYU1OB%2FvwHlGvLtir%2B31V%2ByKZ2si3s5O7h6abJ6MkFXZMhtIOKxrUWlJCerXQLQPXMZoqc1vw%2BYmFVC07Uq21YjYyKNdH&#x26;X-Amz-Signature=1f7a1d92c1d172a43ed356f676a18f11924c89621110ae866753d2dbf5bf8ec0&#x26;X-Amz-SignedHeaders=host&#x26;x-id=GetObject"><figcaption>Rosette logo</figcaption></figure><figure><img src="https://upload.wikimedia.org/wikipedia/commons/thumb/2/20/YouTube_2024.svg/500px-YouTube_2024.svg.png"><figcaption>Youtube Logo</figcaption></figure><table><tbody><tr><td>Sr. No</td><td>Name</td><td>Apples</td></tr><tr><td>1</td><td>Alice</td><td><em><strong>6</strong></em></td></tr><tr><td>2</td><td>John</td><td><em><strong>10</strong></em></td></tr></tbody></table><code class="language-math">x^n + y^n = z^n</code><p class=""><strong>Bold </strong><em><strong>Italic </strong></em><u><em><strong>Underline </strong></em></u><u><del><em><strong>StrikeThrough</strong></em></del></u><u><em><strong> </strong></em></u><code><u><em><strong>Code </strong></em></u></code><span class="notion-yellow-background"><code><u><em><strong>Color</strong></em></u></code></span></p><p class=""></p><ul><li class="">Apples</li></ul><ul><li class="">Oranges</li></ul><ul><li class="">Mangoes</li></ul><p class=""></p><ol><li class="">Lakers</li></ol><ol><li class="">Warriors</li></ol><ol><li class="">Mavericks</li></ol><p class=""></p><pre><code class="language-javascript">console.log("Hello World!");</code></pre><p class=""></p><p class=""><a data-notion-page-mention="126744b6-a7b9-433a-bf8a-78db40eb004e">Untitled</a> </p><p class=""></p><figure><video src="https://www.youtube.com/watch?v=dQw4w9WgXcQ" controls></video><figcaption></figcaption></figure><p class=""></p><details><summary class="">What is love?</summary><div><p class="">Baby don’t hurt me…</p></div></details><p class=""></p><blockquote class=""><em><strong>“One must have chaos in one's self to give birth to a dancing star</strong></em><strong>.”</strong> -Nietzsche</blockquote><p class=""></p>
